refac: migrate indexer filter to indexer_api module
