+++ deps/wazuh_testing/wazuh_testing/end_to_end/vulnerability_detector.py
@@ -23,34 +23,12 @@ import json
 from typing import Dict, List
 
 from wazuh_testing.tools.system import HostManager
-from wazuh_testing.end_to_end.indexer_api import get_indexer_values
+from wazuh_testing.end_to_end.indexer_api import get_indexer_values, create_vulnerability_states_indexer_filter
 from wazuh_testing.end_to_end.regex import REGEX_PATTERNS
 from collections import namedtuple
 
 
 Vulnerability = namedtuple('Vulnerability', ['cve', 'package_name', 'package_version', 'type', 'architecture'])
-
-
-def create_vulnerability_states_indexer_filter(target_agent: str, greater_than_timestamp: str) -> dict:
-    return _create_filter(target_agent, greater_than_timestamp, 'vulnerability.detected_at')
-
-
-def create_alerts_filter(target_agent: str, greater_than_timestamp: str) -> dict:
-    return _create_filter(target_agent, greater_than_timestamp, '@timestamp')
-
-
-def _create_filter(target_agent: str, greater_than_timestamp: str, timestamp_field: str) -> dict:
-    filter = {
-        'bool': {
-            'must': []
-        }
-    }
-    if greater_than_timestamp:
-        filter['bool']['must'].append({'range': {timestamp_field: {'gte': greater_than_timestamp}}})
-    if target_agent:
-        filter['bool']['must'].append({'match': {'agent.name': target_agent}})
-
-    return filter
 
 
 def load_packages_metadata() -> Dict:
@@ -101,7 +79,9 @@ def check_vuln_state_index(host_manager: HostManager, host: str, package: Dict[s
         current_datetime (str): Datetime to filter the vulnerability state index.
     """
     filter = create_vulnerability_states_indexer_filter(host, current_datetime)
-    index_vuln_state_content = get_indexer_values(host_manager, index='wazuh-states-vulnerabilities', filter=filter)['hits']['hits']
+    index_vuln_state_content = get_indexer_values(host_manager,
+                                                  index='wazuh-states-vulnerabilities',
+                                                  filter=filter)['hits']['hits']
     expected_alerts_not_found = []
 
     logging.info(f"Checking vulnerability state index {package}")
