+++ tests/end_to_end/test_vulnerability_detector/test_vulnerability_detector.py
@@ -53,12 +53,12 @@ from wazuh_testing.end_to_end.logs import truncate_remote_host_group_files
 from wazuh_testing.end_to_end.waiters import wait_until_vd_is_updated
 from wazuh_testing.end_to_end.monitoring import generate_monitoring_logs, monitoring_events_multihost
 from wazuh_testing.end_to_end.regex import get_event_regex
-from wazuh_testing.end_to_end.indexer_api import get_indexer_values, delete_index
+from wazuh_testing.end_to_end.indexer_api import get_indexer_values, delete_index, \
+        create_vulnerability_states_indexer_filter, create_alerts_filter
 from wazuh_testing.tools.configuration import load_configuration_template
 from wazuh_testing.tools.system import HostManager
 from wazuh_testing.end_to_end.remote_operations_handler import launch_parallel_operations
-from wazuh_testing.end_to_end.vulnerability_detector import get_vulnerabilities_from_states, \
-        create_vulnerability_states_indexer_filter, create_alerts_filter
+from wazuh_testing.end_to_end.vulnerability_detector import get_vulnerabilities_from_states
 from wazuh_testing.modules.syscollector import TIMEOUT_SYSCOLLECTOR_SCAN
 
 
@@ -333,9 +333,9 @@ class TestInitialScans():
         time.sleep(TIMEOUT_PER_AGENT_VULNERABILITY_SCAN * len(agents_to_check))
 
         for agent in agents_to_check:
-            agent_all_vulnerabilities = get_indexer_values(host_manager, 
-                                                           filter=create_vulnerability_states_indexer_filter(target_agent=agent, 
-                                                                                                             greater_than_timestamp=setup_vulnerability_tests),
+            filter = create_vulnerability_states_indexer_filter(agent, setup_vulnerability_tests)
+            agent_all_vulnerabilities = get_indexer_values(host_manager,
+                                                           filter=filter,
                                                            index='wazuh-states-vulnerabilities')['hits']['hits']
 
             vuln_by_agent_index[agent] = agent_all_vulnerabilities
